requirements.txt
Django>=2.2
djangorestframework
python-magic>=0.4,<0.5
psycopg2
